requirements.txt
opencv-python
customtkinter
Pillow
numpy